refactor: delete and move components or unused files
