+++ src/components/ui/CardResults.tsx
@@ -1,6 +1,6 @@
-import Card from '../Card/Card';
 import { Country } from '../../@types/countries';
 import { Grid } from '@radix-ui/themes';
+import Card from './Card';
 
 interface CardResultsProps {
   countries: Country[];
